Split interface APIs to libvirtApi/interface.js
